+++ ui/src/components/collections/schema/SchemaFieldJson.svelte
@@ -8,17 +8,7 @@
     let showInfo = false;
 </script>
 
-<SchemaField
-    bind:field
-    {key}
-    on:rename
-    on:remove
-    on:drop
-    on:dragstart
-    on:dragenter
-    on:dragleave
-    {...$$restProps}
->
+<SchemaField bind:field {key} on:rename on:remove {...$$restProps}>
     <svelte:fragment slot="options">
         <button
             type="button"